feat: change diary audio field name

+++ controllers/diary.controller.js.js
@@ -6,21 +6,26 @@ const { lightFormat, addDays } = require("date-fns");
 
 const Diary = require("../models/Diary");
 const { User } = require("../models/User");
-const { RESPONSE, DATE_FORMAT, DIARY_SENTIMENT } = require("../constants");
+const {
+  RESPONSE,
+  DATE_FORMAT,
+  DIARY_SENTIMENT,
+  DIARY,
+} = require("../constants");
 const sentiment = new Sentiment();
 
 exports.createDiary = async (req, res, next) => {
   try {
     const { email } = req.user;
-    const { script } = req.body;
-    const audio = req.file.location;
+    const { script = DIARY.DEFAULT_SCRIPT } = req.body;
+    const audioURL = req.file.location;
 
     const sentimentResult = sentiment.analyze(script);
 
     const { _id: userId } = await User.findOne({ email }).select("_id").lean();
 
     const newDiary = {
-      audio,
+      audioURL,
       script,
       sentiment: sentimentResult.score,
       createdAt: new Date().toISOString(),
@@ -149,6 +154,11 @@ exports.getDiary = async (req, res, next) => {
 exports.deleteDiary = async (req, res, next) => {
   try {
     const { diary_id } = req.params;
+
+    if (!mongoose.Types.ObjectId.isValid(diary_id)) {
+      next(createError.BadRequest());
+      return;
+    }
 
     await Diary.findByIdAndDelete(diary_id);
